backend/requirements.txt
fastmcp>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
google-generativeai>=0.3.2  # <-- The correct, official package
python-dotenv>=1.0.0
pyyaml>=6.0
pathlib2>=2.3.0
flask>=2.3.0
requests>=2.31.0
# New dependencies for n8n integration
sqlalchemy>=1.4.0
cryptography>=3.4.0

# Authentication dependencies
flask-login>=0.6.0
authlib>=1.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
psycopg2-binary>=2.9.0